# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Four Delivery API",
        "description": "API para o sistema de delivery Four Delivery",
        "contact": {
            "email": "suporte@fourdelivery.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/admin-v1",
            "description": "Admin API V1"
        },
        {
            "url": "/api/app-v1",
            "description": "App API V1"
        }
    ],
    "paths": {
...
        "/user/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Login de Usuário (Admin)",
                "operationId": "3e02ce1202147650082f27f189ccf07d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login bem-sucedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Authorized"
                                        },
                                        "token": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                }
            }
        },
        "/customer/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Login de Cliente",
                "operationId": "93067db3d786f99fbd7e688fa547071d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "customer@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login bem-sucedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Authorized"
                                        },
                                        "token": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                }
            }
        },
        "/check-token-validity": {
            "get": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Verificar validade do token",
                "operationId": "88800c9736d459bfdacf36199e64b03c",
                "responses": {
                    "200": {
                        "description": "Token válido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "authenticated": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido ou ausente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Logout de Usuário/Cliente",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logout realizado com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/register": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Registrar nova empresa e usuário",
                "operationId": "1f06d3130b898773356c61c7e17c857a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "company_name",
                                    "company_fantasy_name",
                                    "company_email",
                                    "company_phone",
                                    "user_name",
                                    "user_email",
                                    "user_phone",
                                    "user_password",
                                    "user_password_confirmation"
                                ],
                                "properties": {
                                    "company_name": {
                                        "type": "string",
                                        "example": "Empresa Teste"
                                    },
                                    "company_fantasy_name": {
                                        "type": "string",
                                        "example": "Teste"
                                    },
                                    "company_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "contato@empresa.com"
                                    },
                                    "company_phone": {
                                        "type": "string",
                                        "example": "11999999999"
                                    },
                                    "user_name": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "user_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@example.com"
                                    },
                                    "user_phone": {
                                        "type": "string",
                                        "example": "11999999999"
                                    },
                                    "user_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "user_password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registro realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Company and user registered successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados inválidos"
                    },
                    "500": {
                        "description": "Erro interno"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}